==> notebook/01_twitter_API_tweets_tweepy.ipynb <==
## 1. Get Tweets From Twitter API

### Import packages

import pandas as pd
import tweepy as tw
import os
from tqdm import tqdm

# pd.set_option('max_rows', None)

### Twitter API authentication

# Twitter API keys and tokens
API_KEY = os.environ.get("API_KEY")
SECRET_KEY = os.environ.get("SECRET_KEY")
ACCESS_TOKEN = os.environ.get("ACCESS_TOKEN")
SECRET_TOKEN = os.environ.get("SECRET_TOKEN")

# establish an instance of tweepy's API
auth = tw.OAuthHandler(API_KEY, SECRET_KEY)
auth.set_access_token(ACCESS_TOKEN, SECRET_TOKEN)
api = tw.API(auth, wait_on_rate_limit=True)

try:
    api.verify_credentials()

    print("Authentication OK")

except:

    print("Error during authentication")

### Create variables

# Date 
since_date = '2022-05-24'
until_date = '2022-05-25'
filename_date = since_date.replace('-', '_')

# Count
count_en = 10000
count_other = 5000

# Query  
query_ukraine_en = f'ukraine OR ukrainian OR kyiv OR kharkiv OR odessa OR donetsk OR lviv OR mariupol OR luhansk OR donbas OR donbass OR kherson -Zelenskiy -Zelensky -Zelenskyy -@ZelenskyyUa -russia -russian -moscow -putin -@KremlinRussia_E since:{since_date} until:{until_date} -filter:retweets'
query_ukraine_es = f'ucrania OR ucraniano OR kyiv OR kharkiv OR odessa OR donetsk OR lviv OR mariupol OR lugansk OR donbas OR donbass OR kherson -zelenskiy -@ZelenskyyUa -rusia -ruso -moscú -putin -@KremlinRussia_E since:{since_date} until:{until_date} -filter:retweets'
query_ukraine_pt = f'ucrânia OR ucraniano OR kyiv OR kharkiv OR odessa OR donetsk OR lviv OR mariupol OR luhansk OR donbas OR donbass OR kherson -zelenskyy -@ZelenskyyUa -rússia -russo -moscou -putin -@KremlinRussia_E since:{since_date} until:{until_date} -filter:retweets'
query_ukraine_fr = f'ukraine OR ukrainien OR kyiv OR kharkiv OR odessa OR donetsk OR lviv OR mariupol OR luhansk OR donbas OR donbass OR kherson -zelensky -@ZelenskyyUa -russie -russe -moscou -poutine -@KremlinRussia_E since:{since_date} until:{until_date} -filter:retweets'

query_russia_en = f'russia OR russian OR moscow -ukraine -ukrainian -kyiv -Zelenskiy -Zelensky -@ZelenskyyUa -putin -@KremlinRussia_E since:{since_date} until:{until_date} -filter:retweets'
query_russia_es = f'rusia OR ruso OR moscú -ucrania -ucraniano -kyiv -zelenskiy -@ZelenskyyUa -putin -@KremlinRussia_E since:{since_date} until:{until_date} -filter:retweets'
query_russia_pt = f'rússia OR russo OR moscou -ucrânia -ucraniano -kyiv -zelenskyy -@ZelenskyyUa -putin -@KremlinRussia_E since:{since_date} until:{until_date} -filter:retweets'
query_russia_fr = f'russie OR russe OR moscou -ukraine -ukrainien -kyiv -zelensky -@ZelenskyyUa -poutine -@KremlinRussia_E since:{since_date} until:{until_date} -filter:retweets'

query_zelenskyy_en = f'Zelenskiy OR Zelensky OR Zelenskyy OR @ZelenskyyUa -ukraine -ukrainian -kyiv -russia -russian -moscow -putin -@KremlinRussia_E since:{since_date} until:{until_date} -filter:retweets'
query_zelenskyy_es = f'zelenskiy OR @ZelenskyyUa -ucrania -ucraniano -kyiv -rusia -ruso -moscú -putin -@KremlinRussia_E since:{since_date} until:{until_date} -filter:retweets'
query_zelenskyy_pt = f'zelenskyy OR @ZelenskyyUa -ucrânia -ucraniano -kyiv -rússia -russo -moscou -putin -@KremlinRussia_E since:{since_date} until:{until_date} -filter:retweets'
query_zelenskyy_fr = f'zelensky OR @ZelenskyyUa -ukraine -ukrainien -kyiv -russie -russe -moscou -poutine -@KremlinRussia_E since:{since_date} until:{until_date} -filter:retweets'

query_putin_en = f'putin OR @KremlinRussia_E -russia -russian -moscow -ukraine -ukrainian -kyiv -Zelenskiy -Zelensky -Zelenskyy -@ZelenskyyUa since:{since_date} until:{until_date} -filter:retweets'
query_putin_es = f'putin OR @KremlinRussia_E -rusia -ruso -moscú -ucrania -ucraniano -kyiv -zelenskiy -@ZelenskyyUa since:{since_date} until:{until_date} -filter:retweets'
query_putin_pt = f'putin OR @KremlinRussia_E -rússia -russo -moscou -ucrânia -ucraniano -kyiv -zelenskyy -@ZelenskyyUa since:{since_date} until:{until_date} -filter:retweets'
query_putin_fr = f'poutine OR @KremlinRussia_E -russie -russe -moscou -ukraine -ukrainien -kyiv -zelensky -@ZelenskyyUa since:{since_date} until:{until_date} -filter:retweets'

query_nato_en = f'NATO OR "North Atlantic Treaty Organization" OR @NATO since:{since_date} until:{until_date} -filter:retweets'
query_nato_es = f'OTAN OR "Organización del Tratado del Atlántico Norte" OR @NATO since:{since_date} until:{until_date} -filter:retweets'
query_nato_pt = f'OTAN OR "Organização do Tratado do Atlântico Norte" OR @NATO since:{since_date} until:{until_date} -filter:retweets'
query_nato_fr = f'OTAN OR "Organisation du Traité de l\'Atlantique Nord" OR @NATO since:{since_date} until:{until_date} -filter:retweets'

query_eu_en = f'"European Union" OR @EU_Commission since:{since_date} until:{until_date} -filter:retweets'
query_eu_es = f'"Unión Europea" OR @EU_Commission since:{since_date} until:{until_date} -filter:retweets'
query_eu_pt = f'"União Européia" OR @EU_Commission since:{since_date} until:{until_date} -filter:retweets'
query_eu_fr = f'"Union Européenne" OR @EU_Commission since:{since_date} until:{until_date} -filter:retweets'

query_biden = f'"Joe Biden" OR @JoeBiden OR @POTUS since:{since_date} until:{until_date} -filter:retweets'

query_johnson = f'"Boris Johnson" OR @10DowningStreet OR @BorisJohnson since:{since_date} until:{until_date} -filter:retweets'

query_macron = f'"Emmanuel Macron" OR @EmmanuelMacron since:{since_date} until:{until_date} -filter:retweets'

query_scholz = f'"Olaf Scholz" OR @OlafScholz since:{since_date} until:{until_date} -filter:retweets'

### Create functions to get tweets from Twitter API

# get_tweets with encode('utf-8') for en, es, pt, fr
def get_tweets_utf8(query, lang, count, keyword):
    tweets = pd.DataFrame(columns=['tweet_id',  'acct_desc', 'date', 'location', 'friends_ount','followers_ount', 'text'])
    counter = 0
    for tweet in tqdm(tw.Cursor(api.search_tweets, count=count, tweet_mode='extended', q=query, lang=lang).items()): 
            appending_dataframe = pd.DataFrame([[tweet.id, tweet.user.description, tweet.created_at, tweet.user.location, tweet.user.friends_count, tweet.user.followers_count, tweet.full_text.encode('utf-8')]], columns=['tweet_id',  'acct_desc', 'date', 'location', 'friends_ount','followers_ount', 'text'])
            tweets = tweets.append(appending_dataframe)
            counter+=1
            if counter >= count:
                break
    tweets.to_csv(f'../data/tweets/tweets_{keyword}/tweets_{keyword}_{lang}/tweets_{keyword}_{lang}_{filename_date}.csv', index=False)


### Get tweets from Twitter API using Standard Search

# ukraine
get_tweets_utf8(query_ukraine_en, "en", count_en, "ukraine")
get_tweets_utf8(query_ukraine_es, "es", count_other, "ukraine")
get_tweets_utf8(query_ukraine_pt, "pt", count_other, "ukraine")
get_tweets_utf8(query_ukraine_fr, "fr", count_other, "ukraine")

# russia
get_tweets_utf8(query_russia_en, "en", count_en, "russia")
get_tweets_utf8(query_russia_es, "es", count_other, "russia")
get_tweets_utf8(query_russia_pt, "pt", count_other, "russia")
get_tweets_utf8(query_russia_fr, "fr", count_other, "russia")

# zelenskyy
get_tweets_utf8(query_zelenskyy_en, "en", count_en, "zelenskyy")
get_tweets_utf8(query_zelenskyy_es, "es", count_other, "zelenskyy")
get_tweets_utf8(query_zelenskyy_pt, "pt", count_other, "zelenskyy")
get_tweets_utf8(query_zelenskyy_fr, "fr", count_other, "zelenskyy")

# putin
get_tweets_utf8(query_putin_en, "en", count_en, "putin")
get_tweets_utf8(query_putin_es, "es", count_other, "putin")
get_tweets_utf8(query_putin_pt, "pt", count_other, "putin")
get_tweets_utf8(query_putin_fr, "fr", count_other, "putin")

# nato
get_tweets_utf8(query_nato_en, "en", count_en, "nato")
get_tweets_utf8(query_nato_es, "es", count_other, "nato")
get_tweets_utf8(query_nato_pt, "pt", count_other, "nato")
get_tweets_utf8(query_nato_fr, "fr", count_other, "nato")

# eu
get_tweets_utf8(query_eu_en, "en", count_en, "eu")
get_tweets_utf8(query_eu_es, "es", count_other, "eu")
get_tweets_utf8(query_eu_pt, "pt", count_other, "eu")
get_tweets_utf8(query_eu_fr, "fr", count_other, "eu")


# biden
get_tweets_utf8(query_biden, "en", count_en, "biden")
get_tweets_utf8(query_biden, "es", count_other, "biden")
get_tweets_utf8(query_biden, "pt", count_other, "biden")
get_tweets_utf8(query_biden, "fr", count_other, "biden")

# johnson
get_tweets_utf8(query_johnson, "en", count_en, "johnson")
get_tweets_utf8(query_johnson, "es", count_other, "johnson")
get_tweets_utf8(query_johnson, "pt", count_other, "johnson")
get_tweets_utf8(query_johnson, "fr", count_other, "johnson")

# macron
get_tweets_utf8(query_macron, "en", count_en, "macron")
get_tweets_utf8(query_macron, "es", count_other, "macron")
get_tweets_utf8(query_macron, "pt", count_other, "macron")
get_tweets_utf8(query_macron, "fr", count_other, "macron")

# scholz
get_tweets_utf8(query_scholz, "en", count_en, "scholz")
get_tweets_utf8(query_scholz, "es", count_other, "scholz")
get_tweets_utf8(query_scholz, "pt", count_other, "scholz")
get_tweets_utf8(query_scholz, "fr", count_other, "scholz")
