==> notebook/03_VADER_sentiment.ipynb <==
## 3. Sentiment Analysis Using vaderSentiment

### Import packages

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from tqdm.auto import tqdm
import glob
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
tqdm.pandas()

# pd.set_option('display.max_colwidth', None)

### vaderSentiment analysis

analyzer = SentimentIntensityAnalyzer()

# Sentiment analysis of English tweets
for filepath in tqdm(glob.glob('../data/tweets_en/*')):
    df = pd.read_csv(filepath, dtype = {'text':'str'} ,lineterminator='\n', encoding='latin-1')
    df = df.dropna(subset='text')
    compound = []
    for s in df['text'].apply(analyzer.polarity_scores):
        compound.append(s['compound'])
    df['compound'] = compound
    df.to_csv(filepath, index = False)

# Sentiment analysis of non-English tweets
for filepath in tqdm(glob.glob('../data/tweets_translated/*')):
    df = pd.read_csv(filepath, dtype = {'text':'str', 'location_trans':'str'} ,lineterminator='\n', encoding='latin-1')
    df = df.dropna(subset='text')
    compound = []
    for s in df['text'].apply(analyzer.polarity_scores):
        compound.append(s['compound'])
    df['compound'] = compound
    df.to_csv(filepath, index = False)
