==> notebook/06_word_cloud.ipynb <==
## 6. Word Cloud

### Import packages

import numpy as np
import pandas as pd
from os import path
from PIL import Image
from wordcloud import WordCloud, STOPWORDS, ImageColorGenerator
import matplotlib.pyplot as plt
from tqdm.auto import tqdm

### Word cloud of positive tweets

startdate =pd.to_datetime('2022-02-24').date()
keywords = ['ukraine', 'russia', 'eu', 'zelenskyy', 'biden', 'putin', 'johnson', 'nato', 'scholz', 'macron']

for keyword in tqdm(keywords):
    df = pd.read_csv(f'../data/tweets_en/tweets_{keyword}_en.csv',dtype={'date':'str'}, parse_dates = ['date'], lineterminator='\n', encoding='latin-1')
    df['date'] = pd.DatetimeIndex(df['date']).date
    
    # positive tweets
    df = df[(df['date']>startdate) & (df['compound']>0.5)]
    print(keyword, df.shape)
    
    # Concatenate tweets in one string
    text_temp = ''
    for text in df['text']:
        text_temp += text
    
    # Create wordcloud
    wordcloud = WordCloud(background_color="white").generate(text)
#     plt.figure()
#     plt.imshow(wordcloud, interpolation="bilinear")
#     plt.axis("off")
#     plt.show()
    
    # Save the image in the img folder:
    wordcloud.to_file(f"../shinyapp/www/{keyword}_positive.png")  

### Word cloud of negative tweets

for keyword in tqdm(keywords):
    df = pd.read_csv(f'../data/tweets_en/tweets_{keyword}_en.csv',dtype={'date':'str'}, parse_dates = ['date'], lineterminator='\n', encoding='latin-1')
    df['date'] = pd.DatetimeIndex(df['date']).date
    
    # positive tweets
    df = df[(df['date']>startdate) & (df['compound'] < -0.5)]
    print(keyword, df.shape)
    
    # Concatenate tweets in one string
    text_temp = ''
    for text in df['text']:
        text_temp += text
    
    # Create wordcloud
    wordcloud = WordCloud(background_color="white").generate(text)
    
    # Save the image in the img folder:
    wordcloud.to_file(f"../shinyapp/www/{keyword}_negative.png")
